Custom Fonts and error fixes

+++ Maker/src/Game/entity/Entity.java
@@ -27,11 +27,11 @@ public class Entity {
 			removed = true;
 		}
 		
-		public double getX() {
+		public int getX() {
 			return x;
 		}
 		
-		public double getY() {
+		public int getY() {
 			return y;
 		}
